requirements.txt
mysql-connector-python
pandas
tqdm
